+++ src/shared/files.h
@@ -3,6 +3,7 @@
 #include "dynamic_array.h"
 #include "static_array.h"
 #include "strings.h"
+#include "types.h"
 #include <stdio.h>
 #include <stdlib.h>
 
@@ -14,17 +15,17 @@ static inline String read_entire_file(const char *path)
   if (!file) return result;
 
   fseek(file, 0, SEEK_END);
-  long file_size = ftell(file);
+  s64 file_size = (s64)ftell(file);
   fseek(file, 0, SEEK_SET);
 
-  result.items = (char *)malloc(file_size);
+  result.items = (char *)malloc((u64)file_size);
   if (!result.items)
   {
     result = (String){0};
     goto out;
   }
 
-  result.len = fread(result.items, 1, file_size, file);
+  result.len = (u64)fread(result.items, 1, (u64)file_size, file);
 
 out:
   fclose(file);
